Move require check into merge_strict_context
Summary:
Before this diff, check_require called Module_js.module_exists, which is just
NameHeap.mem. We then called NameHeap.get inside of merge_strict_context. This
diff combines those heap lookups.

The missing module error comes from a LookupT. Before this diff, we were
creating a tvar for the output of the lookup, but we never actually used it.
This diff also changes that from a tvar (which will pollute the graph with junk)
to a MixedT.

Reviewed By: mroch

Differential Revision: D5165984

fbshipit-source-id: a1a08fd6f679b9908ea4144b666d5b18e3b4528e