refactor Types_js into some helper methods
Summary:
`full_check` and `recheck` were doing lots of things. this diff splits
them up to make them a bit easier to follow. also, having helpers like `parse`
and `reparse` next to each other makes it easier to compare the two and keep
them in sync when necessary.

this also pulls a bunch of the post-processing that the master does after
calling workers inside the `with_timer` blocks. for example, I think merging the
error maps returned by the "Infer" phase should still count as "Infer", not be
lost in an untimed block.

Reviewed By: samwgoldman

Differential Revision: D5164999

fbshipit-source-id: 8c4d3dddf7fa18ed97123edaa19036ff3e76889c