lazy mode support for check-contents
Summary:
Earlier, check-contents would assume that any dependencies were already checked.

This is not a safe assumption in lazy mode.

This diff triggers checking of dependencies between the "infer" and "merge"
steps of checking the contents of a file. Note that checking might update the
server environment, e.g. with new errors.

(Note: this ignores all push blocking failures!)

Reviewed By: samwgoldman

Differential Revision: D5190274

fbshipit-source-id: d0ec1f133dfa3ab00069b1c5e3f09c0a4382c879