split Types_js.server_init
Summary:
`server_init` does a full check (in fact, it is the only caller of
`full_check`), but a full check involves a bunch of initial work like listing
all of the files, parsing them, and building the dependency graph, before it
does any actual typechecking.

this diff splits `server_init` into `init` and `full_check`.

Reviewed By: gabelevi

Differential Revision: D5203451

fbshipit-source-id: e6cc27c7dbb142d13af3ee06c38b11441f3ed3b0