incremental step to remove Type.t from Flow_error
Summary:
when a `Type.t` and `Type.t` error, the only thing the error message
depends on is their reasons. this adds a special case called `EIncompatibleDefs`
that takes two reasons instead of the full types.

longer term, I want to try to remove all `Type.t`'s from `Flow_error`. this
would save memory in the master, and make it cheaper to ship `Flow_error` across
the socket to the client commands.

Reviewed By: avikchaudhuri

Differential Revision: D5213132

fbshipit-source-id: b84cc6f607ffbf7ee3d5cd8fd8a70d5ccf1e34cc