[hack] Dont retry monitor connection on timeout - ameliorate inadvertent DDOS
Summary:
It has been observed when distributing work across 1000 hh_clients (with, for example, xargs), this can wedge the server. This is because new client connections to the monitor are being created by the retry logic in the client faster than cancelled (due to timeout) connections are being discarded.

I will copy-paste details from the diff itself:

     * The DDOS occurs when the Monitor's new connections queue (arriving on
     * the socket) grows faster than they are being processed. This can happen
     * in two scenarios:
       * 1) Malicious DDOSer fills up new connection queue (incoming
       *    connections on the socket) quicker than the queue is being
       *    consumed.
       * 2) New client connections to the monitor are being created by the
       *    retry logic in hh_client faster than those cancelled connections
       *    (cancelled due to the timeout above) are being discarded by the
       *    monitor. This could happen from thousands of hh_clients being
       *    used to parallelize a job. This is effectively an inadvertent DDOS.
       *    In detail, suppose the timeout above is set to 1 ssecond. 1000
       *    thousand hh_client have timed out at the line above. Then these
       *    1000 clients will cancel the connection and retry. But the Monitor's
       *    connection queue still has these dead/canceled connections waiting
       *    to be processed. Suppose it takes the monitor longer than 1
       *    millisecond to handle and discard a dead connection. Then the
       *    1000 retrying hh_clients will again add another 1000 dead
       *    connections during retrying even tho the monitor has discarded
       *    fewer than 1000 dead connections. Thus, no progress will be made
       *    on clearing out dead connections and all new connection attempts
       *    will time out.
       *
       *    We ameliorate this by having the timeout be quite large
       *    (many seconds) and by not auto-retrying connections to the Monitor.

Note: Minor tradeoff in that we are no longer printing the progress spinner while retrying on Monitor connections (because we no longer retry on Monitor connections). Instead, we're just giving the monitor many seconds to establish a connection, during which we print no spinner/progress bar at all on terminal. This is mostly benign since the Monitor will always be available except under extremely heavy load. We still have a progress bar during the Server portion of the work (after Prehandoff).

Note 2: This prevents 1000 hh_clients from wedging a server. But there is still a bug somewhere where a few of these 1000 hh_clients thinks that Prehandoff has finished even though the actual handoff to the Server (Server's connection queue full).

Differential Revision: D5205759

fbshipit-source-id: 4f445dec9f4b048539f652430a6a994361ab2d89