[hack] Increase max purgatory clients limit
Summary:
A monitor can be running without a server (server status is Informant_killed) while it is
waiting for the informant to start one up (typically awaiting an hg update to complete so we can load a fresh saved state on a new server).

In this situation, the monitor will send Server_not_alive_dormant and then stash the connected File Descriptor
away inside env.purgatory_clients to be processed at some later time (after the Informant starts a server). Because we don't want this queue to be growing unbounded, it has a size limit.

Hitting this limit has actually been observed in the wild (see attached task). It's unclear why so many
client connections would be spawned, but I'm upping this limit for now to ameliorate. Our hh_client doesn't behave badly in this way - when it sees the Server_not_alive_dormant message, it blocks patiently during prehandoff waiting for the next message. I've tested this, and there is no malignent explosion in connection attempts.

Reviewed By: dlreeves

Differential Revision: D5185200

fbshipit-source-id: a6f37583f328683f3cb46b2fba9353a95861aa59