split ESpeculationFailed into union- and intersection-specific variants
Summary:
`ESpeculationFailed` is used for both union and intersection failures.
however, the upper bound passed for unions is always `UseT (UnknownUse, EmptyT)`,
which makes much of the `ESpeculationFailed` error printing a no-op.

by spliting it into `EUnionSpeculationFailed`, we can simplify to the point
where we only need to pass reasons instead of types.

`EIntersectionSpeculationFailed` is harder, since the upper can be anything.
will follow up on that later.

Reviewed By: samwgoldman

Differential Revision: D5233432

fbshipit-source-id: 17d16f7efc8a56f0e529d79ec9689cd1c2dcc2d9