opt for bad sentinel check with enums
Summary:
Consider o.key === v where both o.key and v have very large union types. This is
usually a mistake when used to refine stuff: o.key should have a singleton type
for this check to be useful.

In any case, today this causes a quadratic blowup of constraints generated, all
for nothing.

This diff uses UnionRep.quick_mem to optimize the processing of this check.

Reviewed By: samwgoldman, gabelevi, mroch

Differential Revision: D5253394

fbshipit-source-id: 67e61c20518e95faf2f38d9641d12dc5535c2b3e