Don't cache reads from the ContextHeap
Summary:
During merge, we only read from the ContextHeap once per file, unlike the
SigContextHeap, which may be read once for each time it appears as a dependency.

However, we were caching reads from the heap. Twice, in fact! The
SharedMem.WithCache functor creates a module with an internal per-worker cache.
Because contexts are mutable, we needed to copy the worker-cached contexts and
re-cache them.

Because we only read once, per-worker caching provides no benefit. Since we
don't cache in the worker, there's no reason to copy and re-cache either. We
simply read the context once, mutate it, and throw it away with the worker's
forked process.

Reviewed By: avikchaudhuri

Differential Revision: D5221917

fbshipit-source-id: 916274f5ca848f8d250ce1ecb9bf12759d329a13