Separate missing and unchecked modules during merge
Summary:
Before this change, missing modules and unchecked ones were combined into the
same list and passed to Merge_js. We still look up unchecked modules (in case
there is a declared module) but we don't expect to find one, so the lookup is
non-strict. That meant the lookup for missing modules had to be nonstrict as
well, so we needed to do an extra strict lookup (missing_required_module).

If we split the lists, we can do a non-strict lookup for unchecked modules and a
strict lookup for missing modules instead.

Reviewed By: avikchaudhuri

Differential Revision: D5248039

fbshipit-source-id: f71c1de2434415c53c863662ee96813092dea2ef