Split liberr test into liberr and liberr_parse
Summary:
The liberr test has a type error and a parse error in the libs and a local type
error in a.js. When I move infer into merge, we no longer see local type error
because we don't merge when after lib parse errors.

Splitting this into a test that asserts on the order of lib vs. non-lib errors
and a test that asserts that local errors appear after lib parse errors.

The assertion in the latter test will change in a follow-up diff.

Reviewed By: avikchaudhuri

Differential Revision: D5248186

fbshipit-source-id: aa161297d77b002157cbbf45c0557826c9af066b