Make ROArray type covariant in variance checking
Summary:
We currently treat the type parameter in ReadOnlyArrays as invariant (neutral polarity).
This leads to the following reporting an error:
```
class Foo<+T> {
  good (x: number) : $ReadOnlyArray<T> {
      return [];
  }
}
```

The error reported is that T is in an invariant position and is expected to only occur covariantly.
Since ReadOnlyArrays are read only, it seems that they should be treated covariantly.

Additionally, the following code typechecks, indicating that we treat it covariantly
in __flow:
```
function good(x: $ReadOnlyArray<number>) : $ReadOnlyArray<number | string> {
  return [];
}
```

Reviewed By: samwgoldman

Differential Revision: D5259916

fbshipit-source-id: 8f56b31be267bdb9e9929e379f7e16a6c3553156