mroch merging 59f6d78241c2b729012dd0797cc8e1ffcfdbb1e9 | avik | don't check a class body twice if its sig doesn't contain `this`
Summary:
(1) Polymorphic definitions are checked multiple times, twice for every type
parameter (once bound to empty, once bound to the declared bound).

(2) All class definitions have an implicit type parameter, `this`.

Together, this means that all class definitions are checked at least twice.

It turns out this is wasteful if the class body itself doesn't have an element
whose type contains `this`. (The superclass, on the other hand, is always
implicitly instantiated with `this`, so there's no waste when doing super
checks.)

Moreover, when combined with type inference for class properties with
initializers, the multiple passes can bloat exported types and lead to perf
problems downstream.

This diff cuts this waste, by not checking a class body whose signature doesn't
contain `this`, half of the time...in particular, when `this` is bound to
empty. This choice makes sense because `this` always appears covariantly, even
in the super class.

It is conceivable that a similar optimization can be done for other type
parameters that are "unused," but that's for another day.

Reviewed By: samwgoldman

Differential Revision: D5163542

fbshipit-source-id: 7ee2b894e6fb5e6e8b97f173156c4f8eb4682ea9