remove Type.use_t from EIncompatibleProp
Summary:
splits `EIncompatibleProp` into separate errors for GetProp and SetProp
and Lookup failures, which removes the dependency on `Type.use_t`, since it was
only used to call `err_operation use_t`.

longer term, I will remove the dependency on `Type.t` as well, but need to
figure out a holistic way to deal with all the `err_value` callers first.

this duplicates the error text from `err_operation`, but I'm working on
incrementally killing `err_operation` so the duplication will be short-lived.

Reviewed By: samwgoldman

Differential Revision: D5233051

fbshipit-source-id: 460ceb9f60ee9f5fc9dde3bf8f05ce094652bb28