don't run full_check during --lazy
Summary:
D5268466 moved most of `full_check` into `Server.init`. what's left in
`full_check` is just a call to `typecheck`. turns out it's a no-op when you call
it with `is_lazy_mode` and `parsed:[]`. so, just don't call it at all. this
makes sense because lazy mode is supposed to start the server and do nothing, so
we wouldn't even want `full_check` to do anything.

Reviewed By: gabelevi

Differential Revision: D5273702

fbshipit-source-id: 07c3db45fe50cf9d0313a833ad34464e5c92cb33