Always use an array for jsonrpc args
Summary:
The spec here is a bit of a mess. The vscode-jsonrpc package has the following behavior:

- If there are no arguments, set `params` to `null`
- If there is a single argument, set `params` to that argument
- If there is more than one argument, set `params` to an array of the given arguments

This is really bad, since there is actually no way to distinguish between some scenarios. For example:

- The resulting messages are identical if the caller provides no arguments and if the caller provides a single `null` argument.
- The resulting messages are identical if the caller provides a two (or more) element array as the only argument, and if the caller provides those two or more elements as individual arguments.

Fortunately, while vscode-jsonrpc uses this insane behavior to determine what to output, it happily accepts input if we just set `params` to be a single array with all the arguments as input. So that's what we'll do here.

Reviewed By: mroch

Differential Revision: D5277304

fbshipit-source-id: 53ec7755ed71b0825f9450d606bad7fec558fe86