Context.copy_of_context should also copy type table
Summary:
`Context.copy_of_context` does a mostly-shallow clone of a context.
This appears to be an accident of history. Copying a context and then calling
`Context.clear_intermediates` on the original should not also clear out the
clone.

This is not an audit of all of the parts of `Context`. I care about the type
table, so that is the only thing handled here.

Reviewed By: samwgoldman

Differential Revision: D5279151

fbshipit-source-id: 5a60823efdcd7e84c6de3c740c522d1cfd77a1ce