Update React libdefs for Fiber
Summary:
React Fiber lets you return more types from your render method. Specifically,
you can return an array of elements, or strings, etc.

This change updates the React libdefs in a few ways.

1. Define a React$Node<T> type, which represents all valid return types
2. Update React.render and like to accept React$Node<T>
   (note that createElement still returns a React$Element)
3. Exports a non-parameterized Node type from the react module

I also updated some of the hardcoded logic in Flow that deals with these types.
Specifically, in the createElement custom fun, we check that a stateless
functional component is correct by "calling" it with an expected return type.
That expected return type is now `React$Node<Config>`.

The non-parameterized exports give folks a way of importing types without
depending on the globals. I used this in the tests in a few places.

In projects that use a lot of React, I recommend creating an alias:

```
type ReactNode = React$Node<any>
```

and using that as the return types of render() methods. If you are doing
something like a HOC, or otherwise need to use the parameterized version, you
are free to do so, but people should use the non-parameterized versions
otherwise.

Note that one could equivalently write something like this in every module:

```
import {Node as ReactNode} from 'react';
```

But I think that's too cumbersome to actually do in practice, if you have lots
of React in your codebase, so just make a global alias.

Reviewed By: spicyj

Differential Revision: D5296482

fbshipit-source-id: 80c09dc22015dc1e53743a352091c26eb1aa6050