[hack] Dont reuse subscription socket for watchman get all files queries
Summary:
There is a bug sometimes when invoking get_all_files.

During server init when get_all_files is invoked but there
is actually some payload sitting in the socket due to the watchman
subscription, we end up reading the subscription response instead of
the all_query response. This results in a crash.

This fixes that by keeping the subscription pipe *only* for reading pushed content
due to the subscription. All other queries open a new one-use connection.

Reviewed By: dlreeves

Differential Revision: D5075826

fbshipit-source-id: 26afc5d1339a06bf774d8bc82a6ed83c57eb3d34