[travis] fix version of npm used to publish flow-parser
Summary:
the "script" deploy step uses an ancient version of node and npm (npm1
and node 0.10). we were expecting npm2's .npmrc format.

the deploy step runs in a separate shell from the build/test steps, so
altough we switch versions of node during the build, it doesn't apply during
deploy. this diff will switch to a modern node in the deploy step.

Reviewed By: nmote

Differential Revision: D5266172

fbshipit-source-id: 8d3477a6a3688b8749d4e376c9b98aa1e0392151