[hack] hh_server reports char at autocomplete position
Summary:
For sake of autocomplete, we'll need to know what is the character at the caret while you're doing the autocomplete.

(Previously this used to be done within Nuclide itself, which knows the content of the buffer. Now we want to move the logic inside the LSP server. But hh_client doesn't know the content of the buffer, so the work has to be computed by hh_server).

The way I've implemented it is a trifle inefficient: it involves a linear scan of the document once to find the caret offset, and a second time to find the range to insert the autocomplete marker.

But, frankly, the whole "linear scan" is what's inefficient. If we wanted to improve efficiency then we'd do it on a grander scale by caching line-start offsets.

Reviewed By: arxanas

Differential Revision: D5329528

fbshipit-source-id: d0feeb1955ca4798845741631011978b71f9f463