[hack] Tagging comments
Summary: Running into dependencies on ad hoc invariants that no longer hold in the full fidelity parser, we need to clean up the handling of comments a little. The `docblock_finder`, for example, identifies whether something is a line comment (as opposed to a block or markup comment) by checking that the //length of the Pos it is tagged with is one//. These types of invariants just won't do. This diff cleans that up.

Reviewed By: oulgen

Differential Revision: D5274249

fbshipit-source-id: a846eae70c6c340dffc9d131403d2e8919e92f0a