Fix missing push diagnostics
Summary: It was possible for the recheck bookends to be sent without an accompanying message afterwards reporting *all* diagnostics. This would cause Nuclide to clear out diagnostics. This links them inextricably, so that if the recheck bookends are sent, they must be followed by a message with the complete set of diagnostics.

Reviewed By: avikchaudhuri

Differential Revision: D5348023

fbshipit-source-id: 9b3d1ce8a5e4131fef21e1826bd38177951956df