[hack] Cleaning up over-engineering of Relative_path
Summary: In trying to clean up the positional information used throughout Hack, there's quite a bit of friction due to the over-engineering (i.e. engineering with the foresight of future work that never happened and is no longer planned) of some of the basic types. In this case `Relative_path.t`. Cleaning up a bit in preparation of more sweeping changes to `Pos`

Reviewed By: dlreeves, vladima

Differential Revision: D5274383

fbshipit-source-id: c2352f73f0f4b6f32030736a7dba6d4aec713d56