clean up locs in Declaration_parser
Summary:
builds on top of D5365875 and D5366416, which refactored how variable
declaration locations are tracked.

this diff just moves around where we call `with_loc`, so that we can avoid
calling it in places that don't use the result. it's also nice that
`Declaration_parser.const` returns a `Statement.VariableDeclaration.t` now,
instead of a `(Loc.t * Statement.VariableDeclaration.t)`. The reason `Loc.t`
isn't built into `Statement.VariableDeclaration.t` is because it isn't always
relevant, and this diff parallels that.

Reviewed By: gabelevi

Differential Revision: D5366591

fbshipit-source-id: 85a099a78a330989d24814d33c07ee99b841699f