fix end location of labelled statements
Summary:
replaces a caller of `semicolon_loc`, which gets closing parens wrong,
with a caller of `with_loc`.

as long as `statement` handles parens properly for other types of statements,
this diff will be a no-op, since the tail of a labelled statement is any other
statement.

changing it will unblock getting rid of `semicolon_loc`.

Reviewed By: gabelevi

Differential Revision: D5367336

fbshipit-source-id: 026297706132971955b73612f6bd3d67fc5ad802