Add React component JSX usage tests
Summary:
This diff was originally a part of D5370791.

This adds a test I've been using which aims to cover all the usages of React components in JSX. (Excluding children which is another set of tests.) It is helpful as it both prevents future regressions, and it serves to track some of the bugs I'm looking to fix in Flow's React integration.

Specifically you can see some incorrect output around Flow for exact props and default props.

If you can think of any other ways to use a React component with JSX let me know and we can add it to this test. If you see any unexpected errors that are not related to exact props and default props also let me know.

Reviewed By: samwgoldman

Differential Revision: D5381951

fbshipit-source-id: e3c4abd74fcbb30fea7ab654edd6ea59ebdfd130