Desketch Default Pattern
Summary:
One of the cases that was being detected by the sketch null check lint
before was the pattern of providing a default (empty) value in place
of a maybe value, such as in `function (x: ?string) {x = x || ''; ...}`. This
isn't sketchy when the value on the right matches the potentially falsey value
on the left. This diff keeps the sketchy null check lint rule from firing on
this pattern, to reduce noise.

Reviewed By: samwgoldman

Differential Revision: D5358957

fbshipit-source-id: 68389420959d778b16437d45c59c09c37c1c135d