[PR] [docs] Correct description of subtype relation
Summary:
Having read this carefully, I think there's a typo - since it's `ObjectB` that contains all the keys of `ObjectA`, `ObjectB` is the subtype and `ObjectA` is the supertype.

Of course, being new to Flow, it's also possible I misunderstood something so please correct me if I'm wrong.

(And apologies for the branch name - I edited this quickly in the web UI without downloading the repo, and as far as I can tell there's [still no support](https://github.com/isaacs/github/issues/228) for supplying a custom branch name when doing it that way)
Closes https://github.com/facebook/flow/pull/4223

Differential Revision: D5393643

Pulled By: gabelevi

fbshipit-source-id: e302096ffeb9298442d9416c8cdc5bf50db8c690