[parser] add DeclareInterface AST node
Summary:
adds a `DeclareInterface` ast node to distinguish interfaces with a
`declare` modifier:

```
declare interface Foo {} // DeclareInterface
interface Foo {} // InterfaceDeclaration
```

currently, both are parsed as `InterfaceDeclaration`s, and the locations of both
only include the `interface Foo {}` part. the `declare` is orphaned. this is
bad.

we actually added this node to babel (https://github.com/babel/babel/pull/3230) but didn't do the same for ourselves.

Reviewed By: samwgoldman

Differential Revision: D5394048

fbshipit-source-id: 43cc2f117ff2501ae16041da15848a1ceeea30a7