move protocols into their own directory
Summary:
Moves `ServerProt` into its own library, instead of being part of
"persistent_connections". I thought I needed to do this to break a cycle, but I
don't think that's true. I think this is cleaner, though, under the assumption
that the protocols should be separate from the libraries that use the protocol,
to avoid cycles.

Reviewed By: gabelevi

Differential Revision: D5404136

fbshipit-source-id: ac2aa60f6550441249039e535b4afee06ca1be3f