Opaque Types with reasons
Summary: An implementation of opaque type aliases using reason locations to decide whether or not to expose the underlying type information. When an opaque type occurs in a lower bound, we allow it to expose its type information if its def_loc and loc are the same. When it's in an upperbound, we check to see if the lower bound has the same loc as the opaque type's def_loc. To handle polymorphic opaque types, we flow the lower bound to the upper when the ids are equal and handle substitution as normal.

Reviewed By: samwgoldman

Differential Revision: D5426749

fbshipit-source-id: cc5ef39b1e507517f5e791b6eb161e5c7bd1f08b