[tests] mark a test as "errored" if test.sh exits abnormally
Summary:
`test.sh` is expected to succeed and exit code 0, because commands that expect
failures should use `assert_exit`. if something goes unexpectedly wrong and it
doesn't exit cleanly (for example, we try to move a file that doesn't exist),
then there's something wrong with the test rather than something wrong with the
thing being tested.

in that case, mark the test as "errored" instead of "failed".

i'm hoping this will help us track down why tests are randomly failing in
Travis.

Reviewed By: gabelevi

Differential Revision: D5429568

fbshipit-source-id: d219c04f475d71ae17f04c2900655f006827e710