[tests] add utility to assert exit codes
Summary:
adds `assert_ok` and `assert_errors` utilities for `test.sh` scripts,
so we can verify commands are running properly. this will let us distinguish
between commands that are failing unexpectedly and those that are expected to
fail (e.g. because they return errors so exit code 2).

This is more explicit than relying on what a command prints to stdout showing up
in the `.diff`, since they should be printing to `stderr` on error!

Reviewed By: gabelevi

Differential Revision: D5428202

fbshipit-source-id: a451235df7d49af4546bf1ebf9dce6b411baeb2c