Opaque type aliases with subtyping constraints
Summary: An implementation for subtyping constraints on top of opaque type aliases. When an opaque type is the lower bound of a flow in a file where it's implementation should be hidden, we allow it to be have as if it were the super type given to it when it was declared. Importantly, we do not do this when it is an upper bound, since this would allow implicit conversions to an opaque type alias, which can violate invariants imposed by the opaque type.

Reviewed By: samwgoldman

Differential Revision: D5427273

fbshipit-source-id: fe127e7b2190070e508d59a5026d00f3190bb8eb