Polymorphic Opaque types in libdefs
Summary: Adds support for polymorphic opaque types defined in libdefs. We have to keep track of the type args and polarities for when we compare two OpaqueTs with the same id. Since they may not have an underlying_t, the call to subst may not actually do anything. In this case, we have the type args set aside so that we can still determine if one is a subtype of the other.

Reviewed By: samwgoldman

Differential Revision: D5445290

fbshipit-source-id: 9635d2425094dcd2ebbb8538aa2ae8aebd21fd1e