[tests] additional logging when 'flow start' fails during tests
Summary:
when a test has a custom shell script, we run `flow start` first. if
it fails to start, the test is considered "errored." but since nothing wrote to
the `.out` file yet, the error handler fails trying to `mv` the `.out` file.
also, it will be useful to know what `flow start`'s exit code was when it failed
to start.

Reviewed By: samwgoldman

Differential Revision: D5470905

fbshipit-source-id: 9460c58f2656e039305e563184ef4359b30d38f0