log requests to the server log
Summary:
logs each request to the server log when it starts. I want this to
debug what the server is doing when the unit tests break in travis, but it also
seems generally useful to be able to have, like the httpd access log.

the only hesitation I have is that I think it's appropriate to use
`Hh_logger.info` instead of `Hh_logger.debug`, but `info` is the default log
level and our logs go to a directory that isn't handled by logrotate. A very
long-running server could generate a really large log file given a ton of
autocomplete and get-def calls.

Reviewed By: gabelevi

Differential Revision: D5482957

fbshipit-source-id: a9562ff0fbfd821b71260bc1095cb91b1726b840