always build lexical scope for program
Summary:
A program opens its own lexical scope. Previously its lexical scope was ignored
when ignore_toplevel=true.

Just like every other variable scope, the associated lexical scope for the
program is nested inside the variable scope.

This diff also cleans up the contract of bindings a bit. Bindings are introduced
in reverse, so given two reverse bindings, which are supposed to be nested, we
append them in reverse order and then reverse the result at the end. This works
but it was tricky. Will clean up the bindings API later but for now, this diff
makes the code more obviously correct.

Reviewed By: mroch

Differential Revision: D5369766

fbshipit-source-id: 5f6a1d5bb0d97623c8f8de59d920925913c4b9c5