Persistent Connections Filter Warnings By File (Flow-Side)
Summary:
This is the Flow code of a change to the persistent connection protocol. The
Nuclide code is in D5439741.

Changed the persistent connection protocol to allow clients to track
and untrack files. Warnings are only shown on tracked files. This allows
Nuclide to track only open files, which keeps large numbers of existing
warnings from overwhelming the persistent connection.

Reviewed By: nmote

Differential Revision: D5439642

fbshipit-source-id: 22b7e7af77bc8ad685584a14d467301bb2c399ca