Correctly type React.Component#setState() updater functions
Summary: Updater functions in `setState()` are strongly recommended in React when updating state based on previous state and are an example of the patterns which may be seen in React stateful functional components. However, updater functions are not currently covered by Flow given the semantics of `$Shape`. As functions expressions have no properties Flow interprets them as matching `$Shape<State>` and does not provide inference for the previous state argument or the return type. (It is worth note TypeScript also has this problem.) This diff adds the type for updater functions to the React definition and changes the semantics of `$Shape<T>` so that function definitions are not accepted. It is worth noting that object types that are callable will still match with `$Shape<T>`.

Reviewed By: samwgoldman

Differential Revision: D5370974

fbshipit-source-id: 1ae26c302c5c1550fb3f26e679a32ebe933e1e91