Fix multiple spreads in JSX element props
Summary:
As I was working in `statement.ml` for JSX props, I noticed that spreads were not completely supported. So this diff resolves some of the todos around JSX spreads by basically copying the infrastructure in `mk_object` and modifying it slightly for use with JSX. I preferred the working implementation to the clean implementation. If you want me to refactor `statement.ml` so that `mk_object` and `jsx_mk_props` use a common code instead of duplicated code, let me know how you would like it.

This diff may also change depending on whether or not we want to land D5371223 in its current form. Also, this diff does not allow spreads when a component has exact props. Will this be a problem?

**Edit:**

It seems like there is a bug that is creating some errors in `www`. I put the following description of the bug in a comment, but the comment keeps getting lost, so I’m moving it to the summary. We should discuss a fix for this before landing the diff:

The following is an error, but should it be given that `number` is overwritten later in the type?

```
({
  ...{number: (null: ?number)},
  number: 42,
}: {
  number: number,
});
```

Reviewed By: samwgoldman

Differential Revision: D5371638

fbshipit-source-id: eb86fc881c68385ba88ead7db8441ed0e2df7ae9