Improve lowers_of_tvar calculation
Summary:
Before this change, the implementation was fairly simple. It just unioned up
all the lower bounds of a tvar.

However, those lower bounds might also include unions, and those unions might
include tvars themselves. This diff goes a bit further to unwrap nested unions
and nested tvars.

Also, when I tried to add empty propagation for VoidT ~> OptionalT, we started
to see errors because the empty was leaking into the module sig during context
opt.  This diff fixes that by ignoring empty if it appears in the lowers.

Reviewed By: avikchaudhuri

Differential Revision: D5501407

fbshipit-source-id: 76ad123ca92102b2b5bae4b8c15529b6cd8a14a4