Changes in typing to use AST dependence
Summary:
This diff helps take advantage of AST dependence to address the issue of losing
refinement information sometimes, and asking for a redundant annotation.

Here is an example:
```
function g() {
  var xx : { p : number } | null = { p : 4 };
  if (xx) {
    return function () {
       /* B : invariant x != undefined needed even if x is not modified elsewhere */
       xx.p = 3;
    }
  }
}
```

As is, Flow would need an invariant at line B, but this is redundant due to
the information provided by `if (xx)`.

With this diff, Flow will remember such information and not ask for
superfluous invariants.

Reviewed By: samwgoldman

Differential Revision: D5386225

fbshipit-source-id: eaf29a62a872590256b239013741c9a4a7e3a3f2