[tests] fix cleanup code that was leaving test dirs around
Summary:
the traps that were supposed to be cleaning up all the temp directories
that runtests.sh creates were not correct. from what I can tell, the `EXIT` trap
that was supposed to delete them only ran when you aborted the tests, not on
successful runs. we need to use `RETURN` for that case.

also, the background tasks appear to continue to run after the main runtests.sh
is aborted. I added a top-level trap to kill the workers.

Reviewed By: gabelevi

Differential Revision: D5492507

fbshipit-source-id: fc649406df3c02a6cffb35becf8afcc59c01bf49