replace node before unifying resolved tvars in goto
Summary:
It's important to replace a node after goto constraints when one is unresolved,
per D3802871, but when both are resolved, things are a bit different.

We need to ensure that t1 and t2 are the same type, but that judgment might be
self-referential, causing an nontermination (observed as a stack overflow in
practice).

This diff changes goto such that we equate the tvars before recursive
unification. Thus, if the unification constraint is self-referential, Flow will
find the tvars equal and stop instead of spin in an endless loop.

Reviewed By: avikchaudhuri

Differential Revision: D5540929

fbshipit-source-id: 31d964244eb7ab684dc089e26e8bb09409574de8