Remove convert command
Summary:
This stuff never became usable and the tests have been disabled since forever.
So much new stuff has been added to Flow and TypeScript since this was written,
it would take an enormous effort to get it halfway working again.

If we ever need this code, it will continue to live on in source control.

Reviewed By: mroch, gabelevi

Differential Revision: D5548539

fbshipit-source-id: 5ce53327d509a990ec8b7335479619167518b688