[PR] Add FreeBSD support
Summary:
This set of changes makes it possible to compile `flow` on FreeBSD.

It does require `libinotify` to be install separately, but otherwise seems fine.  I'm not sure how to make the local `libinotify` a dependency through the `opam` build system.

Additionally, the current install instructions in `README.md` specify that `opam init --comp 4.03.0` is sufficient to change versions but on FreeBSD it is necessary to use `opam switch 4.03.0` since the `init` subcommand will not replace the system version (currently 4.02.3 on FreeBSD) which is automatically installed as a dependency of `opam`.
Closes https://github.com/facebook/flow/pull/4433

Reviewed By: samwgoldman

Differential Revision: D5502390

Pulled By: gabelevi

fbshipit-source-id: 06c2310702c6f34980849776cdbe2fb72542baa1