replace EIntersectionIncompatibleWith* with EIncompatible
Summary:
now that `EIntersectionIncompatibleWith{Def,Use}T` are separate errors,
their implementations are identical to `EIncompatible` and `EIncompatibleDefs`,
with the addition of the pretty-generic `branches`, which are a list of nested
errors.

so, this diff moves `branches` into `EIncompatible` and `EIncompatibleDefs` and
gets rid of `EIntersectionINcompatibleWith*`. in the future, we can likely use
the new `extras` to provide additional proof for other errors.

Reviewed By: gabelevi

Differential Revision: D5463970

fbshipit-source-id: da0d48a46c3715818709e12f8692ca0b2284a43f