Move local inference into merge phase, eliminate separate step
Summary:
This diff is a significant architectural change to Flow's inference engine.

Before these changes, inference was split into two separate phases: infer and
merge. During the infer phase, the statements in each file were processed
separately, with their requires and exports left as open tvars. In the merge
phase, we connect the exports of the dependencies to the imports of their
dependents in topological sort order of the cycles in the project.

The separate phase approach was necessary for a long time, because we gathered
the dependency information during infer. To support lazy type checking, avikchaudhuri
built a pre-pass that extracts the dependencies so we don't need to do local
inference before creating the dependency tree.

Note that each file in a component is still locally inferred in a separate
context with tvars for its imports and exports, which are connected (merged)
separately. The only difference here is that the inference happens in lock-step
with merge.

Reviewed By: avikchaudhuri

Differential Revision: D5549208

fbshipit-source-id: ef0f8bf89dc03100832e41d73d4a67ead08fc977