[parser] test parser against test262
Summary:
a test harness for testing the flow parser against test262. this will
make sure that we can parse files that should be parseable and error on syntax
errors. it doesn't confirm that the resulting ASTs are correct, though. for
that, we would need to print the AST and run the tests.

test262_baseline.txt lists all the currently-failing tests so we can track them
over time.

Reviewed By: avikchaudhuri

Differential Revision: D5586145

fbshipit-source-id: 1e468b379ad9165806f32aaa3bcd4058d029eebc