flow-upgrade initial commit.
Summary:
Since we are going to need to run a couple of codemods against all Flow users' code I built a utility to make this process incredibly simple. The idea is that this utility will enable us to make breaking changes while also maintaining goodwill with Flow users.

I should note that I developed this utility independent of the Flow repo, so it doesn't use any of the infrastructure we already have in place for JavaScript projects. Let me know if you have a good way to integrate this into our infrastructure.

The React upgrades that were added target the 0.53.0 release. Of course we can change the release target if it makes sense.

There a few things left that can be done:

- Currently the main script assumes the user is upgrading from 0.52.0 to 0.53.0. We should add command line arguments to make this configurable.
- Allow the `all` option to be configurable.
- Publish workflow. I also want to publish the package under the name `create-flow-upgrade` to work with the Yarn create feature.
- `README.md`.

Reviewed By: gabelevi

Differential Revision: D5527916

fbshipit-source-id: c743b0465d9156d6e364ca5cdace4aec584b7adc